+++ aws/src/main/java/org/apache/iceberg/aws/glue/GlueCatalog.java
@@ -43,6 +43,7 @@ import org.apache.iceberg.exceptions.AlreadyExistsException;
 import org.apache.iceberg.exceptions.NamespaceNotEmptyException;
 import org.apache.iceberg.exceptions.NoSuchNamespaceException;
 import org.apache.iceberg.exceptions.NoSuchTableException;
+import org.apache.iceberg.io.CloseableGroup;
 import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.relocated.com.google.common.annotations.VisibleForTesting;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
@@ -80,6 +81,7 @@ public class GlueCatalog extends BaseMetastoreCatalog implements Closeable, Supp
   private AwsProperties awsProperties;
   private FileIO fileIO;
   private LockManager lockManager;
+  private CloseableGroup closeableGroup;
 
   /**
    * No-arg constructor to load the catalog dynamically.
@@ -119,6 +121,12 @@ public class GlueCatalog extends BaseMetastoreCatalog implements Closeable, Supp
     this.glue = client;
     this.lockManager = lock;
     this.fileIO = io;
+
+    this.closeableGroup = new CloseableGroup();
+    closeableGroup.addCloseable(glue);
+    closeableGroup.addCloseable(lockManager);
+    closeableGroup.addCloseable(fileIO);
+    closeableGroup.setSuppressCloseFailure(true);
   }
 
   private String cleanWarehousePath(String path) {
@@ -415,7 +423,7 @@ public class GlueCatalog extends BaseMetastoreCatalog implements Closeable, Supp
 
   @Override
   public void close() throws IOException {
-    glue.close();
+    closeableGroup.close();
   }
 
   @Override