+++ core/src/main/java/org/apache/iceberg/MetadataTableUtils.java
@@ -30,11 +30,26 @@ public class MetadataTableUtils {
     return MetadataTableType.from(identifier.name()) != null;
   }
 
+  public static Table createMetadataTableInstance(Table table, MetadataTableType type) {
+    if (table instanceof BaseTable) {
+      TableOperations ops = ((BaseTable) table).operations();
+      return createMetadataTableInstance(ops, table, metadataTableName(table.name(), type), type);
+    } else {
+      throw new IllegalArgumentException(String.format(
+          "Cannot create metadata table for table %s: not a base table", table));
+    }
+  }
+
   public static Table createMetadataTableInstance(TableOperations ops,
                                                   String baseTableName,
                                                   String metadataTableName,
                                                   MetadataTableType type) {
     Table baseTable = new BaseTable(ops, baseTableName);
+    return createMetadataTableInstance(ops, baseTable, metadataTableName, type);
+  }
+
+  private static Table createMetadataTableInstance(TableOperations ops, Table baseTable, String metadataTableName,
+                                                   MetadataTableType type) {
     switch (type) {
       case ENTRIES:
         return new ManifestEntriesTable(ops, baseTable, metadataTableName);
@@ -67,5 +82,9 @@ public class MetadataTableUtils {
     String baseTableName = BaseMetastoreCatalog.fullTableName(catalogName, baseTableIdentifier);
     String metadataTableName = BaseMetastoreCatalog.fullTableName(catalogName, metadataTableIdentifier);
     return createMetadataTableInstance(ops, baseTableName, metadataTableName, type);
+  }
+
+  private static String metadataTableName(String tableName, MetadataTableType type) {
+    return tableName + (tableName.contains("/") ? "#" : ".") + type;
   }
 }